Add condition to check for ties
Renamed WinConditions to more appropriate name
which is to check for when the game ends, including
ties.

It is a tie when all of the outer arrays do not
contain an "empty" element. Since the tie checker
is put at the end of the switch statement, any of
the previous Win Conditions would return true
if the game had been a win instead of a tie.