Refactor to minimize number of modules
Previous iteration of code was decent for a console
game of tic-tac-toe. However, upon reflecting on
the new task of having the game be manually played
(instead of automatically on startup with loops),
while also being interactable by the player, the
previous iteration was simply too convoluted.

Therefore, this change aims to "clean up" the
code by firstly taking away much of the functionality
of the console version.

This change is intended to not work anymore, and
but this must be done in order to keep the parts
that worked, while re-adding back the parts that
did not work in the future.